🎨 convert layout to functional component and migrate mothership to it's own class
