+++ web/components/core/Layout/Layout.tsx
@@ -9,10 +9,7 @@ import { LauncherDialog } from "../../lib/LauncherDialog";
 
 // Styles
 import { LayoutStyle } from "./styles.scss";
-import { ChangeEvent, FormEvent, MouseEvent, useEffect } from "react";
-import { Router } from "next/router";
-
-import * as gtag from "../../../lib/gtag";
+import { ChangeEvent, FormEvent, MouseEvent } from "react";
 
 // Begin Component
 // __________________________________________________________________________________________
@@ -48,17 +45,6 @@ export const Layout: React.FunctionComponent<Props> = ({
   shouldFocus,
   children,
 }) => {
-  // Google Tag Manager
-  useEffect(() => {
-    const handleRouteChange = (url: string) => {
-      gtag.pageview(url);
-    };
-    Router.events.on("routeChangeComplete", handleRouteChange);
-    return () => {
-      Router.events.off("routeChangeComplete", handleRouteChange);
-    };
-  }, []);
-
   return (
     <>
       <GlobalStyle />
