+++ web/lib/gtag.ts
@@ -16,12 +16,16 @@ type GoogleTagEvent = {
 
 // https://developers.google.com/analytics/devguides/collection/gtagjs/pages
 export const pageview = (url: string) => {
-  window.gtag("config", GA_TRACKING_ID, {
-    page_path: url,
-  });
+  if (typeof window !== "undefined") {
+    window.gtag("config", GA_TRACKING_ID, {
+      page_path: url,
+    });
+  }
 };
 
 // https://developers.google.com/analytics/devguides/collection/gtagjs/events
 export const event = ({ action, params }: GoogleTagEvent) => {
-  window.gtag("event", action, params);
+  if (typeof window !== "undefined") {
+    window.gtag("event", action, params);
+  }
 };
